using next/lik istead of push
