+++ pages/stories/[id].js
@@ -1,18 +1,15 @@
 import { useRouter } from "next/router";
-import StoryDetailCard from "@/components/StoryDetailCard";
+import StoryDetailCard from "@/components/Story";
 import stories from "@/public/db";
 
 function StoryDetailPage() {
   const router = useRouter();
   const { id } = router.query;
 
-  console.log("stories:", stories);
-  console.log("id:", id);
-
   const story = stories.find((story) => story.id === parseInt(id));
 
   if (!story) {
-    return <div>Story not found</div>;
+    return <p>Story not found</p>;
   }
 
   return <StoryDetailCard story={story} />;
